smart-city-ai-service/requirements.txt
torch>=2.0.0
torchvision>=0.21.0
pillow>=10.0.0
fastapi>=0.109.1
uvicorn>=0.23.0
python-multipart>=0.0.19
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.18
pydantic>=2.0.2